requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-decouple==3.8
PyMySQL==1.1.0
SQLAlchemy==2.0.23
python-dotenv==1.0.0